requirements.txt
backoff==2.2.1
faiss-cpu==1.8.0.post1
litellm==1.51.0
numpy==1.26.4
pandas==2.2.2
sentence-transformers==3.0.1
tiktoken==0.7.0
tqdm==4.66.4
weaviate-client==4.10.2
pinecone==5.4.2
chromadb==0.6.2
qdrant-client==1.12.2